Split handlebars helpers from gulpfile
