+++ Gulpfile.js
@@ -7,7 +7,9 @@ var gulp = require('gulp'),
 	plumber = require('gulp-plumber'),
     rename = require('gulp-rename'),
     sass = require('gulp-sass'),
-    sourcemaps = require('gulp-sourcemaps');
+    sourcemaps = require('gulp-sourcemaps'),
+    handlebarsHelpers = require('./helpers');
+
 
 var paths = {
 	jsonFiles	: './src/pages/**/*.json',
@@ -25,38 +27,7 @@ gulp.task('hbs', function () {
     var options = {
         ignorePartials: true,
         batch : [paths.partials],
-        helpers: {
-            styles: function (arr) {
-                var str = '';
-                arr.forEach(function(css) {
-                    str += '<link href="' + css.url + '" rel="stylesheet" type="text/css" />';
-                });
-                return new handlebars.Handlebars.SafeString(str);
-            },
-            scripts: function(arr) {
-                var str = '';
-                arr.forEach(function(js) {
-                    str += '<script src="' + js.url + '" type="text/javascript"></script>';
-                });
-                return new handlebars.Handlebars.SafeString(str);
-            },
-            modules: function(arr) {
-                var str = '';
-                arr.forEach(function(module) {
-                    str +=  str += '<link href="' + module.name + '" rel="stylesheet" type="text/css" />';
-                    str +=  str += '<link href="' + module.name + '.medium.css" rel="stylesheet" type="text/css" />';
-                    str +=  str += '<link href="' + module.name + '.large.css" rel="stylesheet" type="text/css" />';
-                });
-                return new handlebars.Handlebars.SafeString(str);
-            },
-            moduleScripts: function(arr) {
-                var str = '';
-                arr.forEach(function(js) {
-                    str += '<script src="js/' + js.name + '.js" type="text/javascript"></script>';
-                });
-                return new handlebars.Handlebars.SafeString(str);
-            }
-        }       
+        helpers: handlebarsHelpers    
     };
     
     return gulp.src(paths.pages)
